+++ test-projects/access-control/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/cypress-project-access-control
+
 # @keystone-alpha/cypress-project-access-control
 
 ## 1.2.8